requirements.txt
Django==3.2.23
django-extensions==3.2.1
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipython==7.34.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.6.0
pytz==2023.3
sentry-sdk==1.22.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
celery==5.3.0
redis==3.4.1
django-celery-beat==2.6.0